refactor: move binary io to util/binary_io
Loader and Observer used to have read_bytes_as and write_as_bytes,
respectively, and those are moved to util.
