+++ mjolnir/core/ObserverBase.hpp
@@ -2,6 +2,7 @@
 #define MJOLNIR_CORE_OBSERVER_BASE_HPP
 #include <mjolnir/core/System.hpp>
 #include <mjolnir/core/ForceField.hpp>
+#include <mjolnir/util/binary_io.hpp>
 #include <string>
 
 namespace mjolnir
@@ -44,18 +45,6 @@ extern template class ObserverBase<SimulatorTraits<float,  UnlimitedBoundary>   
 extern template class ObserverBase<SimulatorTraits<double, CuboidalPeriodicBoundary>>;
 extern template class ObserverBase<SimulatorTraits<float,  CuboidalPeriodicBoundary>>;
 #endif
-
-namespace detail
-{
-// it is a helper function to write value as an array of bytes
-template<typename T>
-void write_as_bytes(std::ostream& os, const T& v) noexcept
-{
-    using Type = typename std::remove_reference<T>::type;
-    os.write(reinterpret_cast<const char*>(std::addressof(v)), sizeof(Type));
-    return;
-}
-} // detail
 
 } // mjolnir
 #endif// MJOLNIR_CORE_OBSERVER_BASE_HPP
